Изменил работу с шаблонами
