+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "webpack --mode production",
-    "start": "webpack serve"
+    "start": "webpack serve",
+    "prettier": "prettier --write src/**/*.js"
   },
   "repository": {
     "type": "git",
@@ -20,15 +21,15 @@
   },
   "homepage": "https://github.com/frontend-park-mail-ru/2025_1_SuperChips#readme",
   "devDependencies": {
+    "html-webpack-plugin": "^5.6.3",
+    "webpack-dev-server": "^5.2.0",
+    "webpack-cli": "^6.0.1",
+    "webpack": "^5.98.0",
     "clean-webpack-plugin": "^4.0.0",
     "copy-webpack-plugin": "^13.0.0",
-    "css-loader": "^7.1.2",
     "handlebars-loader": "^1.7.3",
-    "html-webpack-plugin": "^5.6.3",
     "style-loader": "^4.0.0",
-    "webpack": "^5.98.0",
-    "webpack-cli": "^6.0.1",
-    "webpack-dev-server": "^5.2.0"
+    "css-loader": "^7.1.2"
   },
   "dependencies": {
     "express": "^4.21.2",
