+++ server/server.js
@@ -2,14 +2,6 @@ const express = require('express');
 const path = require('path');
 const app = express();
 
-
-// app.use(express.static(path.join(__dirname, '../public')));
-// app.use(express.static(path.join(__dirname, '../src')));
-//
-// app.get('*', (req, res) => {
-// 	res.sendFile(path.join(__dirname, '../public/index.html'));
-// });
-
 app.use(express.static(path.join(__dirname, 'dist')));
 
 app.get('*', (req, res) => {
