+++ src/app/common.css
@@ -5,67 +5,3 @@
     font-family: FuturaPT, sans-serif;
     font-size: 16px;
 }
-
-.hidden {
-    visibility: hidden;
-}
-
-.main-page {
-    background: linear-gradient(to bottom, #141414, #252525) fixed;
-    display: grid;
-    grid-template-columns: 92px auto;
-    grid-template-rows: 112px auto;
-    grid-template-areas:
-            "header header"
-            "sidebar page";
-    gap: 0;
-}
-
-.signup-page, .login-page {
-    background: linear-gradient(to bottom, #141414, #303030);
-    display: flex;
-    height: 100vh;
-    position: relative;
-    overflow: hidden;
-}
-
-.button {
-    width: 440px;
-    height: 44px;
-    color: white;
-    background-color: black;
-    opacity: 25%;
-    cursor: pointer;
-    border: none;
-    border-radius: 20px;
-    font-size: 20px;
-    margin-top: 30px;
-}
-
-.button:hover {
-    opacity: 60%;
-}
-
-.redirect:hover {
-    cursor: pointer;
-}
-
-.page-logo {
-    margin: 62px 10px 20px 94px;
-    opacity: 80%;
-    width: 300px;
-    height: 160px;
-    -webkit-user-drag: none;
-}
-
-.background-vector {
-    position: absolute;
-    top: 0;
-    left: 0;
-    transform: rotate(110deg) translateY(20%) translateX(15%);
-    -webkit-user-drag: none;
-}
-
-.background {
-    min-width: 0;
-}
