+++ src/pages/feed/ui/feed.js
@@ -1,6 +1,6 @@
+import { createSkeletonFeed } from "../lib/skeleton";
 import { createNavbar } from "../../../widgets/navbar/navbar.js";
 import { createSidebar } from "../../../widgets/sidebar/sidebar.js";
-import { goToPage } from "../../../shared/router.js";
 import './feed.css'
 
 /**
@@ -17,30 +17,10 @@ export const renderFeed = () => {
 	const sidebar = createSidebar();
 	page.appendChild(sidebar);
 
-	const feed = document.createElement('div');
+	const feed = createSkeletonFeed(100);
 	feed.classList.add('feed');
 	feed.id = 'feed';
 	page.appendChild(feed);
-
-	const logout = page.querySelector('#logout');
-	logout.addEventListener('click', (event) => {
-		event.preventDefault();
-		goToPage('login');
-	});
-
-
-	const classes = ['skeleton-1', 'skeleton-2', 'skeleton-3'];
-	function randInt(min, max) {
-		return Math.floor(Math.random() * (max-min+1))
-	}
-
-	for (let i=0; i<100; i++) {
-		const elem = document.createElement('div');
-		const rand = classes[randInt(0, classes.length - 1)];
-		elem.classList.add(rand);
-
-		feed.appendChild(elem);
-	}
 
 	return page;
 }
