+++ src/pages/login/ui/login.css
@@ -1,3 +1,11 @@
+.login-page {
+    background: linear-gradient(to bottom, #141414, #303030);
+    display: flex;
+    height: 100vh;
+    position: relative;
+    overflow: hidden;
+}
+
 .login-container {
     background-color: white;
     display: grid;
@@ -28,9 +36,4 @@
     gap: 4px;
     margin-top:12px;
     font-size: 20px;
-}
-
-.error-message {
-    justify-self: right;
-    height: 15px;
 }
