+++ src/shared/components/authPage/authPage.css
@@ -1,0 +1,50 @@
+
+.error-message {
+    justify-self: right;
+    height: 15px;
+}
+
+.button {
+    width: 440px;
+    height: 44px;
+    color: white;
+    background-color: black;
+    opacity: 25%;
+    cursor: pointer;
+    border: none;
+    border-radius: 20px;
+    font-size: 20px;
+    margin-top: 30px;
+}
+
+.button:hover {
+    opacity: 60%;
+}
+
+.redirect:hover {
+    cursor: pointer;
+}
+
+.page-logo {
+    margin: 62px 10px 20px 94px;
+    opacity: 80%;
+    width: 300px;
+    height: 160px;
+    -webkit-user-drag: none;
+}
+
+.background-vector {
+    position: absolute;
+    top: 0;
+    left: 0;
+    transform: rotate(110deg) translateY(20%) translateX(15%);
+    -webkit-user-drag: none;
+}
+
+.background {
+    min-width: 0;
+}
+
+input[type="date"]::-webkit-calendar-picker-indicator {
+    display: none;
+}
