+++ src/shared/components/authPage/authPage.hbs
@@ -10,7 +10,7 @@
 
             <form class="{{page}}-form">
                 {{#each inputs}}
-                    {{> ../../../shared/components/input/input this}}
+                    <div class="input-placeholder"></div>
                 {{/each}}
 
                 <button type="submit" class="button">{{submitBtn}}</button>
