+++ src/shared/components/input/input.css
@@ -48,3 +48,7 @@
     width: 24px;
     height: 24px;
 }
+
+.hidden {
+    visibility: hidden;
+}
