+++ src/shared/config/constants.js
@@ -1,0 +1,3 @@
+export const constants = {
+	apiBaseUrl: '',
+};
